Key/Value Store fixes (#176)
* Key/Value Store fixes

Key/Value Store fixes and missing features.

* Enlarged channel sizes to avoid blocking socket
* Implemented GetKeys()
* Simplified the watcher to async enumeration only

* Key/Value get with revision

* KV delete and update

* Test fixes

* KV create

* KV history and purge

* KV purge deletes

+++ src/NATS.Client.JetStream/Models/StreamMsgGetRequest.cs
@@ -11,7 +11,7 @@ public record StreamMsgGetRequest
     /// </summary>
     [System.Text.Json.Serialization.JsonPropertyName("seq")]
     [System.Text.Json.Serialization.JsonIgnore(Condition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingDefault)]
-    public int Seq { get; set; } = default!;
+    public ulong Seq { get; set; } = default!;
 
     /// <summary>
     /// Retrieves the last message for a given subject, cannot be combined with seq
