+++ tests/NATS.Client.CheckNativeAot/Program.cs
@@ -99,7 +99,7 @@ async Task JetStreamTests()
         var ack = await js.PublishAsync("events.foo", new TestData { Test = 1 }, opts: new NatsPubOpts { Serializer = TestDataJsonSerializer.Default }, cancellationToken: cts1.Token);
         Assert.Null(ack.Error);
         Assert.Equal("events", ack.Stream);
-        Assert.Equal(1, ack.Seq);
+        Assert.Equal(1, (int)ack.Seq);
         Assert.False(ack.Duplicate);
 
         // Message ID
@@ -111,7 +111,7 @@ async Task JetStreamTests()
             cancellationToken: cts1.Token);
         Assert.Null(ack.Error);
         Assert.Equal("events", ack.Stream);
-        Assert.Equal(2, ack.Seq);
+        Assert.Equal(2, (int)ack.Seq);
         Assert.False(ack.Duplicate);
 
         // Duplicate
@@ -123,7 +123,7 @@ async Task JetStreamTests()
             cancellationToken: cts1.Token);
         Assert.Null(ack.Error);
         Assert.Equal("events", ack.Stream);
-        Assert.Equal(2, ack.Seq);
+        Assert.Equal(2, (int)ack.Seq);
         Assert.True(ack.Duplicate);
 
         // Consume
